+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Redesign queue panel
 - Add queue panel slide-out transition
 - Add `Ctrl+Tab` and `Ctrl+Shift+Tab` to select next/previous playlist
+- Add `Back` and `Forward` shortcuts
 - Make time & volume sliders look ok
 - Small design updates to tracklist header & filter textbox
 - Respect user's locale time